refactor: Modularize access token decoding (#3)
