+++ src/user/utils.py
@@ -4,8 +4,6 @@ from fastapi.responses import HTMLResponse
 from starlette import status
 import jwt
 
-from sqlalchemy.orm import Session
-
 import string
 import random
 import datetime
@@ -15,7 +13,6 @@ from core.redis_config import redis_config
 
 from .models import User
 from .schemas import TokenResponse
-from core.exceptions import BaseCustomException
 
 import smtplib
 from email.mime.multipart import MIMEMultipart
@@ -53,70 +50,6 @@ def generate_jwt(user: User):
     rd.set(f'{user.user_id}_refresh', refresh_token, ex=60 * 60 * 24 * 7)
 
     return TokenResponse(access_token=access_token, refresh_token=refresh_token, user_id=user.user_id)
-
-
-# Decode authorization token
-def decode_authorization_token(authorization: str):
-    # Check if token exists
-    if not authorization:
-        raise ValueError("Token required.")
-
-    # Check bearer
-    if not authorization.startswith('Bearer '):
-        raise ValueError("Invalid token.")
-
-    token = authorization[len('Bearer '):]
-    return token
-
-
-# Decode token
-def decode_token(db: Session, token: str, is_access: bool = True):
-    if is_access:
-        token_type = 'access_token'
-    else:
-        token_type = 'refresh_token'
-
-    # Read payload
-    try:
-        payload = jwt.decode(token, jwt_secret, algorithms=['HS256'])
-    except jwt.ExpiredSignatureError:
-        raise BaseCustomException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Token expired."
-        )
-    except jwt.InvalidTokenError:
-        raise BaseCustomException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Invalid token."
-        )
-
-    # Check type
-    if payload['type'] != token_type:
-        raise BaseCustomException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Invalid token."
-        )
-
-    # Find user from DB
-    user = db.query(User).filter(User.user_id == payload['user_id']).first()
-
-    # Check payload and user
-    if not user:
-        raise BaseCustomException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Authorization failed."
-        )
-
-    # Check Redis (Session)
-    rd = redis_config()
-    rd_refresh = rd.get(f'{user.user_id}_refresh')
-    if not rd_refresh:
-        raise BaseCustomException(
-            status_code=status.HTTP_403_FORBIDDEN,
-            detail="Login required."
-        )
-
-    return user
 
 
 # Generate random code
